+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===================
 - Add link to open results page for closed pharmacies
 - Move map and services links from below opening times to below address
+- Remove unused styles from nearest open results view
 - Update npm dependencies
 - Misc refactorings